Move map features to NavigationFragment
