+++ app/src/main/res/values-v21/dimens.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="button_margin_narrow">16dp</dimen>
+</resources>
