+++ app/src/main/res/values-v21/style_navigation_button.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="NavigationButton">
+        <item name="android:layout_margin">@dimen/button_margin_narrow</item>
+    </style>
+</resources>
