+++ app/src/main/res/values/dimens.xml
@@ -7,6 +7,8 @@
     <dimen name="activity_vertical_margin">16dp</dimen>
     <dimen name="fab_margin">16dp</dimen>
 
+    <dimen name="button_margin_narrow">8dp</dimen>
+
     <!-- Millis margin -->
     <dimen name="millis_margin">10dp</dimen>
