+++ app/src/main/res/values/style_navigation_button.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="NavigationButton">
+        <item name="android:layout_marginLeft">@dimen/button_margin_narrow</item>
+        <item name="android:layout_marginTop">@dimen/button_margin_narrow</item>
+        <item name="android:layout_marginRight">@dimen/button_margin_narrow</item>
+        <item name="android:layout_marginBottom">0dp</item>
+    </style>
+</resources>
